secure auth and support Named-UUID, drop magic string as incompatible to FastAPI

+++ main.py
@@ -1,9 +1,19 @@
 # import os, load dotenv, get SECRET_KEY from .env
 import os
 from dotenv import load_dotenv
+
+from utils.uuid_handling import generate_uuid
 load_dotenv()
 SECRET_KEY = os.getenv("SECRET_KEY")
 SECRET_KEY_BYTES = SECRET_KEY.encode()
+
+DEBUGMODE = int(os.getenv("DEBUGMODE", 0))
+
+from utils.auth import hash_new_password, is_correct_password, serialize_bytes_to_str, deserialize_str_to_bytes
+
+from utils.uuid_handling import generate_uuid, generate_prefixed_uuid, match_prefixed_uuid
+
+from utils.patch_css import patch_markdown_font_size
 
 from nicegui import ui, app
 from fastapi import Request
@@ -15,20 +25,32 @@ import hmac
 import json
 import base64
 
-import uuid
 import shutil
-
-def generate_uuid():
-    return str(uuid.uuid4())
 
 from glob import glob
 
-# HMAC SHA256
-def hmac_sha256(msg, key=SECRET_KEY_BYTES):
-    return hmac.new(key, msg, hashlib.sha256).hexdigest()
+print("All imports successful")
+print("Debug mode:", DEBUGMODE)
+
+app.storage.general.indent = True
+app.storage.general.backup()
+for user_storage in app.storage._users.values():
+    user_storage.indent = True
+    user_storage.backup()
+
+# assert that the factories folder exists, and that all subfolders are named with factory uuid
+if not os.path.exists("factories"):
+    os.makedirs("factories")
+for factory in glob("factories/*"):
+    if not match_prefixed_uuid("factory", os.path.basename(factory)):
+        # rename the folder
+        print("Renaming", factory)
+        new_path = os.path.join(os.path.dirname(factory), generate_prefixed_uuid("factory"))
+        os.rename(factory, new_path)
 
 # in reality users passwords would obviously need to be hashed
-passwords = {'3dpm': '3dprintersarefun!'}
+passwords = {'3dpm': '3dprintersarefun!', 'admin': 'topsecret'}
+priviledged_users = {'admin'}
 unrestricted_page_routes = {'/login', '/reauth'}
 
 
@@ -41,7 +63,6 @@ class AuthMiddleware(BaseHTTPMiddleware):
     async def dispatch(self, request: Request, call_next):
         if not app.storage.user.get('authenticated', False):
             if not request.url.path.startswith('/_nicegui') and request.url.path not in unrestricted_page_routes:
-                app.storage.user['referrer_path'] = request.url.path  # remember where the user wanted to go
                 return RedirectResponse('/login')
         return await call_next(request)
 
@@ -50,92 +71,76 @@ app.add_middleware(AuthMiddleware)
 @ui.page('/login')
 def login():
     def try_login() -> None:  # local function to avoid passing username and password as arguments
-        if passwords.get(username.value) == password.value:
+        if passwords.get(master_username.value) == master_password.value:
+            # add the username and password to app.storage.general['user_pw'], if not exist create dict
+            if 'user_pw' not in app.storage.general:
+                app.storage.general['user_pw'] = {}
+
+            # check if the username already exists
+            if username.value in app.storage.general['user_pw']:
+                ui.notify('Username already exists', color='negative')
+                return
+            
+            # if the account was created with a master username-password set that is in the priviledged_users set, admin=True
+            is_admin = master_username.value in priviledged_users
+            # hash the password and store the salt and hash
+            salt, pw_hash = hash_new_password(password.value)
+            app.storage.general['user_pw'][username.value] = {'salt': serialize_bytes_to_str(salt), 'pw_hash': serialize_bytes_to_str(pw_hash), 'admin': is_admin}
             app.storage.user.update({'username': username.value, 'authenticated': True})
-            ui.navigate.to(app.storage.user.get('referrer_path', '/'))  # go back to where the user wanted to go
+            ui.navigate.to("/")
         else:
             ui.notify('Wrong username or password', color='negative')
 
-    if app.storage.user.get('authenticated', False):
-        return RedirectResponse('/')
+    def try_login_2() -> None:  # local function to avoid passing username and password as arguments
+        # first check if the username exists
+        if not app.storage.general.get('user_pw', {}).get(username_2.value):
+            ui.notify('Wrong username or password', color='negative')
+        else:
+            # get the salt and hash from the storage
+            salt = deserialize_str_to_bytes(app.storage.general['user_pw'][username_2.value]['salt'])
+            pw_hash = deserialize_str_to_bytes(app.storage.general['user_pw'][username_2.value]['pw_hash'])
+            # check if the password is correct
+            if is_correct_password(salt, pw_hash, password_2.value):
+                app.storage.user.update({'username': username_2.value, 'authenticated': True})
+                ui.navigate.to("/")
+            else:
+                ui.notify('Wrong username or password', color='negative')
+
     with ui.card().classes('absolute-center'):
-        username = ui.input('Username').on('keydown.enter', try_login)
-        password = ui.input('Password', password=True, password_toggle_button=True).on('keydown.enter', try_login)
+        ui.label("Create a new account").classes('text-2xl')
+        master_username = ui.input('Master Username').classes("w-full")
+        master_username.on('keydown.enter', try_login)
+        master_password = ui.input('Master Password', password=True, password_toggle_button=True).classes("w-full")
+        master_password.on('keydown.enter', try_login)
+        if DEBUGMODE:
+            # populate the master username and password
+            a_set_of_values = list(passwords.items())[0]
+            master_username.value = a_set_of_values[0]
+            master_password.value = a_set_of_values[1]
+        username = ui.input('Username').classes("w-full")
+        username.on('keydown.enter', try_login)
+        password = ui.input('Password', password=True, password_toggle_button=True).classes("w-full")
+        password.on('keydown.enter', try_login)
         ui.button('Log in', on_click=try_login)
 
-        def handle_magic():
-            ui.navigate.to(f'/reauth?magic_string={magic_string.value}')
-
-        magic_string = ui.input('Magic string').on('keydown.enter', handle_magic)
-        ui.button("Reauthenticate using Magic String", on_click=handle_magic)
-
-@ui.page('/reauth')
-def reauth(magic_string: str):
-
-    # decode the magic string
-    decoded_magic = base64.b64decode(magic_string).decode()
-    dict_magic = json.loads(decoded_magic)
-    id = dict_magic['id']
-    reauthenticate = dict_magic['reauthenticate']
-
-    if hmac_sha256(id.encode()) == reauthenticate:
-        # set the browser id
-        app.storage.browser['id'] = id
-        app.storage.user.update({'authenticated': True})
-        return RedirectResponse('/')
-    
-    else:
-        return RedirectResponse('/login')
-
-@ui.page('/upload_test')
-def main_page():
-    ui.label("Upload test").classes('text-2xl')
-
-    # submit a file
-    ui.upload(on_upload=lambda e: process_file(e)).classes('max-w-full')
-
-    def process_file(e):
-        print(e) # UploadEventArguments(sender=<nicegui.elements.upload.Upload object at 0x00000214AAB4E270>, client=<nicegui.client.Client object at 0x00000214AAB4DD90>, content=<tempfile.SpooledTemporaryFile object at 0x00000214AA921AE0>, name='justcard.pdf', type='application/pdf')
-        # name is accessed by e.name
-        # type is accessed by e.type
-        content = e.content
-        print("Got file", content)
-        
-        # save the content to a file
-        with open("file", "wb") as f:
-            f.write(content.read())
-
-def generate_reauthentication_magic_string():
-    # generate a link that the user can click to reauthenticate
-    dict_magic = {'id': app.storage.browser['id'], 'reauthenticate': hmac_sha256(app.storage.browser['id'].encode())}
-
-    # base64 encode the dict json dump
-    return base64.b64encode(json.dumps(dict_magic).encode()).decode()
+        ui.label("Login").classes('text-2xl')
+        username_2 = ui.input('Username').classes("w-full")
+        username_2.on('keydown.enter', try_login_2)
+        password_2 = ui.input('Password', password=True, password_toggle_button=True).classes("w-full")
+        password_2.on('keydown.enter', try_login_2)
+        ui.button('Log in', on_click=try_login_2)
 
 @ui.page('/')
 def main_page():
     ui.label("Hello, world!").classes('text-2xl')
-    # user id at app.storage.browser['id']
-    ui.label(f"User ID: {app.storage.browser['id']}")
-    
-    ui.label("Token-based authentication").classes('text-xl')
-
-    reauth = generate_reauthentication_magic_string()
-
-    async def copy_to_clipboard():
-        await ui.run_javascript(f'navigator.clipboard.writeText("{reauth}")')
-        ui.notify('Copied to clipboard', color='positive')
-
-    ui.label("If you want to retain your session, copy this link (never expires, so keep it secret):")
-    ui.button("Click and copy URL", on_click=copy_to_clipboard)
-
-    """ui.label("If you want to ")
-    # button to navigate to the upload page
-    ui.button('Upload test', on_click=lambda: ui.navigate.to('/upload_test'))"""
+    username = app.storage.user.get('username', 'not logged in?')
+    ui.label(f"Logged in as {username}")
+    is_admin = app.storage.general.get('user_pw', {}).get(username, {}).get('admin', False)
+    admin_text = "Admin" if is_admin else "Not Admin"
+    ui.label(admin_text)
 
     # button to navigate to the factories page
     ui.button('Factories', on_click=lambda: ui.navigate.to('/show_factories'))
-
 
 @ui.page("/show_factories")
 def show_factories():
@@ -161,15 +166,22 @@ def show_factories():
                     ui.label(f"Factory: {desc['name']}").classes('text-xl')
                     ui.markdown(desc['description'])
                     ui.markdown(desc['upload_instructions'])
-                    ui.image(f"{factory}/{desc['cover_image']}").classes("w-full h-96").props("fit='contain'")
+                    ui.image(f"{factory}/{desc['cover_image']}").classes("w-full h-96").props("fit='contain' ratio=1")
                     ui.button("New Job", on_click=lambda factory=basename: new_job_to_factory(factory))
                     ui.button("Show Jobs", on_click=lambda factory=basename: show_jobs_at_factory(factory))
 
 @ui.page("/new_job_to_factory/{factory}")
 def new_job_to_factory(factory: str):
+    # assert that the factory exists
+    if not match_prefixed_uuid("factory", factory):
+        return ui.navigate.to("/show_factories")
+    if not os.path.exists(f"factories/{factory}/desc.json"):
+        return ui.navigate.to("/show_factories")
+
     # creates /jobs/{random UUID} folder
     # writes the job_info.json file
-    my_uuid = generate_uuid()
+    # my_uuid = generate_uuid()
+    my_uuid = generate_prefixed_uuid("job")
     os.makedirs(f"jobs/{my_uuid}")
     with open(f"jobs/{my_uuid}/job_info.json", "w") as f:
         json.dump({'factory': factory, 'status': 'new'}, f)
@@ -178,6 +190,11 @@ def new_job_to_factory(factory: str):
 
 @ui.page("/submit_to_factory/{job_uuid}")
 def submit_to_factory(job_uuid: str):
+    # assert that the job_uuid exists
+    if not match_prefixed_uuid("job", job_uuid):
+        return ui.navigate.to("/")
+    if not os.path.exists(f"jobs/{job_uuid}/job_info.json"):
+        return ui.navigate.to("/")
 
     all_fields_ready = True
